Add files via upload
The Assessment_Q1 involves looking for;

1. High-Value Customers with Multiple Products
Scenario: The business wants to identify customers who have both a savings and an investment plan (cross-selling opportunity).
Task: Write a query to find customers with at least one funded savings plan AND one funded investment plan, sorted by total deposits.
Tables:
● users_customuser
● savings_savingsaccount
● plans_plan

My approach to this task was first to import the SQL Database (db) provided, then i began querying the db in a new sql query tab named 'Assessment_Q1'

Next was to SHOW the databases (db) on my MySQL server and USE the "adashi_staging" that was provided which i had already imported.

Next was i showed all tables in the db and described them to help me understand their contents using SHOW TABLES and DESCRIBE e.g plans_plan.

Next, spotlighted the required tables for this task which are 3.

One of the difficulties i had was joining the required tables as the ID provided are non-aggregated. I was able to solve this problem by using id which was a primary key in a particular table and a foreign key to another. e.g the owner_id was a primary key in plans_plan table and a foreign key in users_customuser.

Another difficulty i faced was how to separate the ''SAVINGS" from "INVESTMENT" accounts as both where located in the plan_type_id. I was able to solve this using the "DISTINCT CASE WHEN" function for Savings accounts = 1 and Investment accounts = (2, 3, 4)

I have also included in my sql file line by line explanations of complex queries used.